simulation time double > int

+++ src/Simulator.java
@@ -4,12 +4,12 @@ import java.util.*;
 class Simulator {
     final private Queue<Task> tasks = new PriorityQueue<>(new TaskComparator());
     final private HashMap<Integer, List<Task>> cycleTasks;
-    final private double SIMULATION_TIME;
+    final private int SIMULATION_TIME;
     final private Queue<Processor> idleProcessors;
     final private ArrayList<Processor> busyProcessors;
     final private Schedular scheduler;
 
-    public Simulator(int numberOfProcessors, double simulationTime, String tasksFilePath) {
+    public Simulator(int numberOfProcessors, int simulationTime, String tasksFilePath) {
         this.SIMULATION_TIME = simulationTime;
         this.idleProcessors = new LinkedList<>();
         this.busyProcessors = new ArrayList<>();
